Remove AwakeMonitor
Awake monitor is only necessary for UniTask due to how
Unity Monobehaviours functioned. Specifically,
MonoBehaviours would not call OnDestroy if they
were not awaked first. Therefore UniTask had to rely
on checking if the MonoBehaviour was set to null to
detect if the MonoBehavior was destroyed.

However in Godot, Predelete is always called when a node is freed, so there's
no need to handle any edge cases.

See https://github.com/Cysharp/UniTask/issues/435 the use of AwakeMonitor in UniTask.